Move comparer to SchemaGeneratorContext
- Comparer class has been moved from ObjectSchemaGenerator to SchemaGeneratorContext.
- SchemaGeneratorContext has a lazy-initialised property that will create a comparer for the context's type. This will allow caching of the comparer instance.
